+++ src/test/java/seedu/address/logic/parser/parserutil/ParserUtilTest.java
@@ -1,0 +1,216 @@
+package seedu.address.logic.parser.parserutil;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
+
+import static seedu.address.logic.parser.ParserUtil.MESSAGE_INVALID_INDEX;
+import static seedu.address.logic.parser.ParserUtil.parseEmail;
+import static seedu.address.logic.parser.ParserUtil.parsePhone;
+import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
+
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.Optional;
+import java.util.Set;
+
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.rules.ExpectedException;
+
+import seedu.address.commons.exceptions.IllegalValueException;
+import seedu.address.logic.parser.ParserUtil;
+import seedu.address.model.person.Address;
+import seedu.address.model.person.Email;
+import seedu.address.model.person.Name;
+import seedu.address.model.person.Phone;
+import seedu.address.model.person.Remark;
+import seedu.address.model.tag.Tag;
+
+public class ParserUtilTest {
+    private static final String INVALID_NAME = "R@chel";
+    private static final String INVALID_PHONE = "+651234";
+    private static final String INVALID_ADDRESS = " ";
+    private static final String INVALID_EMAIL = "example.com";
+    private static final String INVALID_TAG = "#friend";
+
+    private static final String VALID_NAME = "Rachel Walker";
+    private static final String VALID_PHONE = "1234567";
+    private static final String VALID_ADDRESS = "123 Main Street #0505";
+    private static final String VALID_EMAIL = "rachel@example.com";
+    private static final String VALID_REMARK = "Likes to eat.";
+    private static final String VALID_TAG_1 = "friend";
+    private static final String VALID_TAG_2 = "neighbour";
+
+    @Rule
+    public final ExpectedException thrown = ExpectedException.none();
+
+    @Test
+    public void parseIndexInvalidInputThrowsIllegalValueException() throws Exception {
+        thrown.expect(IllegalValueException.class);
+        ParserUtil.parseIndex("10 a");
+    }
+
+    @Test
+    public void parseIndexOutOfRangeInputThrowsIllegalValueException() throws Exception {
+        thrown.expect(IllegalValueException.class);
+        thrown.expectMessage(MESSAGE_INVALID_INDEX);
+        ParserUtil.parseIndex(Long.toString(Integer.MAX_VALUE + 1));
+    }
+
+    @Test
+    public void parseIndexValidInputSuccess() throws Exception {
+        // No whitespaces
+        assertEquals(INDEX_FIRST_PERSON, ParserUtil.parseIndex("1"));
+
+        // Leading and trailing whitespaces
+        assertEquals(INDEX_FIRST_PERSON, ParserUtil.parseIndex("  1  "));
+    }
+
+    @Test
+    public void parseNameNullThrowsNullPointerException() throws Exception {
+        thrown.expect(NullPointerException.class);
+        ParserUtil.parseName(null);
+    }
+
+    @Test
+    public void parseNameInvalidValueThrowsIllegalValueException() throws Exception {
+        thrown.expect(IllegalValueException.class);
+        ParserUtil.parseName(Optional.of(INVALID_NAME));
+    }
+
+    @Test
+    public void parseNameOptionalEmptyReturnsOptionalEmpty() throws Exception {
+        assertFalse(ParserUtil.parseName(Optional.empty()).isPresent());
+    }
+
+    @Test
+    public void parseNameValidValueReturnsName() throws Exception {
+        Name expectedName = new Name(VALID_NAME);
+        Optional<Name> actualName = ParserUtil.parseName(Optional.of(VALID_NAME));
+
+        assertEquals(expectedName, actualName.get());
+    }
+
+    @Test
+    public void parsePhoneNullThrowsNullPointerException() throws Exception {
+        thrown.expect(NullPointerException.class);
+        parsePhone(null);
+    }
+
+    @Test
+    public void parsePhoneInvalidValueThrowsIllegalValueException() throws Exception {
+        thrown.expect(IllegalValueException.class);
+        parsePhone(Optional.of(INVALID_PHONE));
+    }
+
+    @Test
+    public void parsePhoneOptionalEmptyReturnsOptionalEmpty() throws Exception {
+        assertFalse(parsePhone(Optional.empty()).isPresent());
+    }
+
+    @Test
+    public void parsePhoneValidValueReturnsPhone() throws Exception {
+        Phone expectedPhone = new Phone(VALID_PHONE);
+        Optional<Phone> actualPhone = parsePhone(Optional.of(VALID_PHONE));
+
+        assertEquals(expectedPhone, actualPhone.get());
+    }
+
+    @Test
+    public void parseAddressNullThrowsNullPointerException() throws Exception {
+        thrown.expect(NullPointerException.class);
+        ParserUtil.parseAddress(null);
+    }
+
+    @Test
+    public void parseAddressInvalidValueThrowsIllegalValueException() throws Exception {
+        thrown.expect(IllegalValueException.class);
+        ParserUtil.parseAddress(Optional.of(INVALID_ADDRESS));
+    }
+
+    @Test
+    public void parseAddressOptionalEmptyReturnsOptionalEmpty() throws Exception {
+        assertFalse(ParserUtil.parseAddress(Optional.empty()).isPresent());
+    }
+
+    @Test
+    public void parseAddressValidValueReturnsAddress() throws Exception {
+        Address expectedAddress = new Address(VALID_ADDRESS);
+        Optional<Address> actualAddress = ParserUtil.parseAddress(Optional.of(VALID_ADDRESS));
+
+        assertEquals(expectedAddress, actualAddress.get());
+    }
+
+    @Test
+    public void parseEmailNullThrowsNullPointerException() throws Exception {
+        thrown.expect(NullPointerException.class);
+        parseEmail(null);
+    }
+
+    @Test
+    public void parseEmailInvalidValueThrowsIllegalValueException() throws Exception {
+        thrown.expect(IllegalValueException.class);
+        parseEmail(Optional.of(INVALID_EMAIL));
+    }
+
+    @Test
+    public void parseEmailOptionalEmptyReturnsOptionalEmpty() throws Exception {
+        assertFalse(parseEmail(Optional.empty()).isPresent());
+    }
+
+    @Test
+    public void parseEmailValidValueReturnsEmail() throws Exception {
+        Email expectedEmail = new Email(VALID_EMAIL);
+        Optional<Email> actualEmail = parseEmail(Optional.of(VALID_EMAIL));
+
+        assertEquals(expectedEmail, actualEmail.get());
+    }
+
+    @Test
+    public void parseRemarkNullThrowsNullPointerException() throws Exception {
+        thrown.expect(NullPointerException.class);
+        ParserUtil.parseRemark(null);
+    }
+
+    @Test
+    public void parseRemarkOptionalEmptyReturnsOptionalEmpty() throws Exception {
+        assertFalse(ParserUtil.parseRemark(Optional.empty()).isPresent());
+    }
+
+    @Test
+    public void parseRemarkValidValueReturnsRemark() throws Exception {
+        Remark expectedRemark = new Remark(VALID_REMARK);
+        Optional<Remark> actualRemark = ParserUtil.parseRemark(Optional.of(VALID_REMARK));
+
+        assertEquals(expectedRemark, actualRemark.get());
+    }
+
+
+
+    @Test
+    public void parseTagsNullThrowsNullPointerException() throws Exception {
+        thrown.expect(NullPointerException.class);
+        ParserUtil.parseTags(null);
+    }
+
+    @Test
+    public void parseTagsCollectionWithInvalidTagsThrowsIllegalValueException() throws Exception {
+        thrown.expect(IllegalValueException.class);
+        ParserUtil.parseTags(Arrays.asList(VALID_TAG_1, INVALID_TAG));
+    }
+
+    @Test
+    public void parseTagsEmptyCollectionReturnsEmptySet() throws Exception {
+        assertTrue(ParserUtil.parseTags(Collections.emptyList()).isEmpty());
+    }
+
+    @Test
+    public void parseTagsCollectionWithValidTagsReturnsTagSet() throws Exception {
+        Set<Tag> actualTagSet = ParserUtil.parseTags(Arrays.asList(VALID_TAG_1, VALID_TAG_2));
+        Set<Tag> expectedTagSet = new HashSet<Tag>(Arrays.asList(new Tag(VALID_TAG_1), new Tag(VALID_TAG_2)));
+
+        assertEquals(expectedTagSet, actualTagSet);
+    }
+}
